Override board bounding box from generic JSON based on edges
